Add reverse geocoding when adding new trees
I am using the iOS CLGeocoder but I put the code
in OTMAPI so that if a future version of the 
app requires a custom reverse geocoder, the
implementation can be swapped in with minimal
impact.

The documentation for CLGeocoder recommended only
having one reverse geocode operation running at
a time, so I implemented the method as as an NSOperation that I add to an operation queue with
the max concurrent requests set to 1.