Add a progress spinner to the find location button
This shows the user that a geolocation is in
progress. To make this work I am swapping out the
'pointer' image with a transparent one and then
floating an activity view over the button.

The only way to shrink the size of the button to
match the '+' button was to shrink the pointer icon.