Add expiration to AZMemoryObjectCache
The map tiles that are being generated on the
iOS device and cached need to expire so that they
will be regenerated with fresh data from the server
and show trees added or removed by other users.

I implemented this in the generic 
AZMemoryObjectCache by wrapping the actual objects
being cached with a new AZCachedObject class that
records the time. When objectForKey: is called,
the age of the object is checked and the object
is purged if it is too old.