Add a "mini-map" to the tree detail view
I am adding a small map to the detail view to
solve two problems.

- When arriving on the tree details page after
clicking an item in the nearby trees list, the
map gives the user context about the location
of the tree.

- In a future commit, clicking the mini map 
while the tree details page is in edit mode will
allow the user to adjust the position of the tree.