Add the ability to edit plot/tree location
- Added the ability to click on the mini map in
edit mode to slide in a map view on which you can
drag the point to edit the location.

- Added a "detail.png" overlay to the mini map
when the tree detail page is in edit mode to make
it look like the other table cells that slide in
new views when they are clicked.

- If the tree location is changed cached tile
images are invalidated and the map is refreshed.