Add a field detail view for showing pending edits
I added a new view controller that slides in when
selecting a field on the tree detail view. For
ordinary users, this will explain what a pending
field is and show the "real" value. For admins,
this will be the view where pending edits can be
approved or rejected.

The field detail view is a contentless shell at
this point.