Show scientific name on tree on specifies select
The tree detail view was not updating the
scientific name when a new species was selected.
After exposing the scientific name to the view it
was only slightly more effort to show it in in the
list and on the edit mode species cell.

During the update I also replaced creation of the
species list cells in code with a prototype cell
in the storyboard, for easier visual tweaking.