Update field configuration to support ownership
The tree.species field 'owns' the tree.sci_name
and tree.species_name fields. Being able to set
an owner in the config allows the view controllers
to check for pending edits on a fields's owner
and show the appropriate version of a field.

Proper functioning of this feature is dependent on 
the API including a dictionary of related_fields 
for each pending edit.