Style the field edit reject button and add confirmation
The reject all button was a little too close to the
other UI elements so I now show the action sheet
as confirmation to guard against accidental taps.

The SDK does not provide a helper for creating the
big red delete button used to indicate a destructive action so I build the button manually
with an image background.

I downloaded the red glassy button image from
this stack overflow answer
http://stackoverflow.com/a/4261426