Update pending edit approval/rejection to support location
I reused the map view cell from the tree detail
view to show the original and pending edits. It
may be cleaner to show them on a full screen map 
view as multiple annotation but the UI would have 
been harder to implement and there has already been
a lot of time spent on the pending feature, which
is used by a minority of implementations.