Fix crash when updating plot without tree
The stripPendingImageData method was failing if
the data NSDictionary contained a "tree" item set
to <null>

We have discussed fixing the API interface so that
keys with <null> values are simply removed. This
would eliminate this class of bug, but needs a
to be refactored with care so that any code relying
on <null> values is updated.