Fix pending edit bug when returning to map
After editing a plot/tree, the edits were not
being sent back to the map view. Going from
tree detail to the map view then immediately back
to the detail view would cause stale data to be
set on the detail view.

I made use of the existing 
viewController:editedTree:withOriginalLocation:
method to sync the edits with the map view.

BugzID: 24420