Fix tree edit to store empty values as nil not ""
Before this commit, placing the focus on a field
like Tree Height, leaving it empty, and then
trying to save would raise an exception. The
exception was caused by the fact that the client
was sending { tree: { height: "" } } and the
empty string failed to be converted into a number.

My solution was to update the edit view table
cell to store nil rather than "" if no value
is entered.

BugzID: 24411