Fix memory exception when root VC was out of scope
When a seque is "swapped" out as the root VC, nothing maintains
a reference to the view controller.

Since we only use this in one place (the initial delay), a
forced retain was inserted.

If there is a better way to chagne the root vc for the app, it
should supersede this commit