Add support for loading about.html
This commit adds support for loading a custom
about.html page in a UIWebView on the About tab,
allowing for an arbitrary amount of scrolling text
and links.

The project references 3 new files that _must_ be
included in every skin to ensure that the project
builds correctly.

  - about.html
  - about_header@2x.png
  - about_footer@2x.png

If the about.html file is less than 64 bytes
(to allow for an explanatory comment) the
UIWebView is hidden and about_bg.png is shown
as in previous version of the app.