Add in new filter type for choices-based lists
Previously we just had generic "Boolean" and "Range" filters.
This commit adds a new filter type:
  OTMChoiceFilter

If the filter type is selected the dictionary should also
provide a "OTMChoiceFilterChoice" key with the name of the
choice field that they wish to filter filter on.

A complete filter definition looks something like this:
<key>OTMFilterKeyKey</key>
<string>filter_pests</string>
<key>OTMFilterKeyName</key>
<string>Pests</string>
<key>OTMFilterKeyType</key>
<string>OTMChoiceFilter</string>
<key>OTMChoiceFilterChoice</key>
<string>pests</string>