Don't allow users to edit readonly trees
OTM supports a 'readonly' flag which is exposed by the api as
"perm.plot.can_edit", however OTM-iOS was ignoring this flag.

Now we disable the edit button if the flag is set. We also force a
tree/plot update when the user logs in (since the default is that they
can't really edit anything).