Allow editing if no permission data given
Previously we did this by checking the "perms" section of the tree data.
If that section didn't exist we assumed they had editing power.

If a person only creates a plot the perms section will exist but wont
contain any information about the tree. This commit now assumes that
missing tree or plot information should be treated as 'edit allowed'