Preload user info
Right now the whole app is in somewhat of a flux due to the way user and
instance loading works.

Right now we make two async requests right away:

(1) Get instance details
(2) Check user's login

Only when (2) completes does the login manager make the user
available. This means that (1) is always unauthenticated - which means
it doesn't work.

This commit is a bit of duct tape to allow us to work around this for
the moment until we make the instance manager