Fix nearby tree table cells
The table cells were expecting the OTM1 formatted JSON dictionaries and
were, therefore, displaying nulls for all items.

I temporarily disabled displaying the distance from the current location
to each tree because the unit of measurement needs to come from the
instance configuration, which is a larger feature that I don't want to
tackle right now.