Load branding preferences from API endpoint
The instance API response now includes the "config" dictionary as well
as the the custom logo url. I have added code to set these values on the
environment and raise a notification that the environment changes. The
various UI widgets respond to the notification and change their colors
accordingly.

I removed navBarTintColor from the environment and replaced it with
primaryColor (and secondary color) to avoid confusion. The color may be
used for more than just nav bar tinting.