Update unit label handling
The formatter was incorrectly writing the unit label into the text box
when switching to edit mode so I added a label to the table cells that
will display the unit next to the text field.

I also assume that if a field has a numeric formatter, that the decimal
keyboard should be shown rather than the full-text keyboard.