Update new photo thumbnail on map detail view
Tree photos are added from the tree detail view but were not being
pushed back to the map view. When returning to the map view from the
detail view, the stale image gave the appearance that the photo upload
failed.

This implementation is a little complicated, due to the existing
complexity of the async photo upload workflow. It could use a refactor,
but we are trying to get a release out the door.