Reference local images in Xcode project
There was a sneaky problem in Jenkins where skin assets were being
pulled from the wrong folder. The SD app was showing LA's default tree
image. The root of the problem appears to be that the project file
referenced a symlink path rather than a local path.

I removed the images folder from the project, copied one of the skin
image folder into the local project path, and then added the images back
into the project. In my local testing, I could now succesfully replace
the real directory with a symlink and the images would correctly change
inside of Xcode.