Update the species edit row with species name.
We want the user to know that his/her choice was respected when editing the tree
details and changing the speices. To do this we add the current species to the
table cell and update it when a new species is selected. This requires making
the tree detail view a delegate for data coming from selected rows on the
species table view controller.