Add section labels to tree edit screen.
We want users to know what the fields they are editing are related to in
reference to the tree detail screen. We had labels for the sections on the
detail screen but not the edit screen so we add them here by reshuffling the
same labels that are used on the edit screen.