Wire the new instance switcher to exisiting elements.
We need the user flow of the application to include the instance switcher. If a
user first starts the app, they will be directed to choose and instance and if
they use the profile page, they will have the option to change locations as
well. The choice of instance is stored in the user's preferences.