Species selection returns scientific name.
The species selection already returned a common name to the delegate but did not
return the scientific name. To provide more flexability for the future we add
the scientific name as well. Where we use this value and update the cell in the
tree detail view, we split out the magic string update code to a separeate
method on the OTMTreeDetailViewController to make it easier to deal with
updating this cell in other places in the code.