Allow apps to be non-instance-switchable.
We need certain apps to remain dedicated solely to a certain instance. In these
cases, the app must not be allowed to switch to other instances. We had
overlooked a scenario in which the user could access the instance switcher from
the profile screen of apps that were meant to be single instance. To address
this we use a boolean set on the OTM shared environment settings. The value of
this boolean is determined based on the value of the "instance" string. Empty or
missing strings make the instance switchabe. The controllers will reference this
logic and add or hide the buttons from the screen as appropriate.