Add new table view controller for all instances.
This new controller calls the shared settings which updates the list of
instances and then generates an alphabetically sorted list of table cells. The
cells are added to the controller in a single section. The all maps view  has a
delegate that will be set to the instance switcher controller so that controller
can update perform the instance switch which avoids the instance switch logic in
multiple places in the code.