Check for duplicate email message.
When creating a new user, the api server responds with a 409 when there is a
confilct for an email address or username. This was causing confution because
the app did not check which error was encountered and alerted the user that the
username was duplicated when sometimes it was the email address. In order to fix
this we check for the string returned by the api server in the message body and
use this to set a different error message. This is not wonderful since it relies
on the exact string from the API but I am adding a warning in the api server
code to let other devs know that changing that string might have bad
implications.