Restyle registration page.
Our registration page had some styling issues. Here we add in the profile image
from the web site as a placeholder. This is done by layering the user added
image on top of a static image that is controlled in the login storyboard.
Background color is also added to the view controller and autocorrect is
disabled for username, first name, and last name fields.