Adjust element sizes and positions.
The filter button on the filter page was being pushed up off the page in certain
instances after added a speices. I moved the button off of the subview and onto
the parent view and rearranged the sizes accordingly. I also made the search
filter a weak element since it shouldn't persist after the view controller is
removed.