Switch instance list to occur on viewDidLoad.
In testing I ran into an issue with the server and it caused the all instances
list to fail. Because we were dealing with animated views and were loading data
and causing failures before the view was actually finished, the application
would crash. Moving the fetching of the instances, and associated error handling
later (after the view is done loading) prevents this issue. I also removed
commented out code from the recently created controller.