Dismiss the login overlay for logged in users.
We were seeing issues in which users who logged in from a screen that was not
the profile screen would see the login screen even though they were logged in.
This was caused by the controller checking to see if it had ever presented the
user with the login page. We dismiss that screen for users who have logged in
status.