Make sure to reset profile data in edge case.
Found an edge case in which a user goes to profile while logged out, dismisses
the modal, then goes to main map, then logs in through the add tree screen then
returns to the profile page. This caused profile data not to be set. Here we
trap that edge condition and reset the data.