Make detail cell able to display date picker.
Detail cells may need to display date data and use a date picker interface to
update. We want to do this because we have a date planted field that is
identified in the api response for fields on some instances as a date. This
requires a format of YYYY-MM-DD (ex 2014-12-31) but had a plain text field that
would cause api errors and prevent users from saving new or edited trees. We
override the delegate for the normal text field and replace the keyboard with
a date picker. If the picker is set on the object we prevent normal text field
behavior and alert the listening delegate of the update from the date picker
target/action listener.