Fields use date pickers based on api response.
The instance endpoint returns data for fields including their type. This was not
being used to actually alter the interface in the app. Here we watch for "date"
type fields and pass that information to the cell renderers which build out the
detail cells. If a field is a date type we set the cell to use a date picker for
input and formatting rather than a keyboard.