Don't allow editing non-editable fields.
One of our instances had species set as non-editable for public users. We want
to prevent this from causing save errors. Here we look for species and photo
permissions and set them in the shared state. We alter the cells on the tree
detail view to remove click handling if permission is denied. This should get
reset each time the user or instance changes.