Set date picker start date to set value.
When tapping a date field the default start date had been set for the current
day. This is fine if there is no start date however if there is an existing date
the picker should start with that date set so that users can make small
alterations from the current value.