Use array of dictionaries for cell & section data.
We had been using two arrays to deal with tracking the data for section headers
and cells. We would add and remove data from these and so had to continue to
update code with tweaks to point to the proper index for a section title or
specific cell. This caused lots of problems and confusion. Instead we build up
a mutable array of dictionaries. Each dictionary has a "title" which will be the
section title and "cells" which is a mutable array of the cells in the section.
This way the single structure can be responsible for rendering the entire table
and adding a cell does not require adding or subtracting from indexes with
magical numbers. To reduce complexity we also re-render the table when
transitioning to and from edit mode. This is also done after login to avoid
stale data.