Don't rely on cell placement to update cell text.
When swapping the label of the set species cell we had been relying on the
knowledge that it was the first index in the list of visible cells. This
resulted in hard coding a magic number.

The original solution was also flawed because updating the visible cell in the
table is not a sustainable solution. Cells get re-rendered when they  leave the
view via scrolling and thus the text would revert. It was also possible to
update the text in the wrong cell by shifting the topmost cell off the screen
and then editing. This udpates the default text on the renderer and then
re-renders the cell based on indexpath thus ensuring the proper cell text
changes and remains changed.