Create 'holder' class to handle rendered cells.
We need to abstract our cell handling up a level. We had been using a
cell renderer which could return a rendered cell. This assumed a 1:1
relationship between the renderer and a cell (i.e. each renderer could
generate a single cell with data). The addition of stewardship and
alerts means this is no longer true. A single field can now result in
the creation of many cells in the tree deail table. A stewarship field
might have 10 stewarship actions. Additionally these can be intermingled
with a totally different field type (tree stewardships and plot
stewardships). So we need a way to render a group of cells from the
data and move them around in such a way that they still have useful data
on them that can be used for sorting and grouping. Here we make a
CellSorter which is a container that holds individual cells along with a
sort key that can be used to group cells of common type and sort data
that can be used to perform the actual sorting.