Group & sort common fields on tree detail table.
In order to show grouped stewardship and alerts properly we need to find
grouped fields and sort them together by the sort data as defined in the
sort key of our new cell sorter holder. In this case, all the cells need
to be prerendered and sorted before we can actually display them on the
table. This allows us to stash away overflow cells on the Load More
button and re-add them to the table each time the button is clicked.
When the load more button has no more hidden cells on it, it will
dissapear.