Add new stewadships & alerts from tree edit page.
When on the tree edit page we want add a button that allows users to add
a new stewardship or alert UDF. Users can add as many of these as they
would like. The information for the steps in the UDF is set at the time
the data for the instance comes from the API. We create an add more
button for each UDF field and then group those buttons together,
squashing them and combining their field keys for the first step. This
allow one button to handle several similar kinds of UDFs that are part
of the same collection.