Allow saving UDF data on new trees.
The initial data structure of new trees prevented UDF collection data
from being properly set on them. Here we capture the incoming collection
data and use it to set the proper dictionary keys so that we can add a
UDF collection to a new tree at the time we are adding it.