Allow for editing existing alerts.
UDF collections can be edited if they have a step that has the "default"
key. If they do not, they will show but will not have a click callback.
Right now this is only true for alerts. If a UDF has the default key for
one of its steps, that step will be the only editable value. We presume
that there can only be one default. This also means that we need to set
a temporary id on new alerts so that we can edit those without making a
new one. The temp id must be removed prior to saving the data.

Since we need to find the updated UDF and swap out its data prior to
saving, we can't just pass back a single value, we need the entire data
object for the UDF, this means that we need to pass that into the
controller and the reset it on the tree model. To do this we provide
another paramter on the click callback of the cell which allows the
existing data to be sent in and altered.