Auto set the default case on new alerts.
Alerts provide a data key that species their default value. The case for
this is that alerts should always start as unresolved and then later be
marked as resolved. In order to deal with this, we inspect the incoming
data and then hop over the step in the UDF that has a default value and
set the default for the new UDF.