Adding human labels to add more cells.
We need to translate the UDF keys to human readable labels so
tree.udf:Alert needs to be "Tree" and plot.udf:Alert needs to be
"Planting Site." We create an array that matches the incoming UDF key
choices in length and uses a dictionary to set human readable values.
Using the two arrays and their indexes it is easy to know which machine
value to pass to the API and which label to present to the user.