Add filter class for 'Default' type data.
We have creted a new type of field called a 'default' field. These are
fields that have a default value but are otherwise just a choice list.
This was developed to handle alerts which default to an unresolved
state. Since these are UDFs we need to pass in the id of the udf to the
tiler along with the data. Thus we look up the default value when
creating a new default filter and pass it in when creating the object.

Default type filters are also OR type filters for the tiler. This way a
user can search for plots and trees that have open alerts and get both
in the retured data set not just trees that have alerts and alerts for
their planting sites. In order to do this we had to restructure the data
passed in to the tiler query and create a sub-array that holds the OR
type filters.