Use constants for date strings.
Since our date strings are used in many places, it will be annoying if
an API change forces us to go through many files altering strings to
deal with an update. Instead we set a constant on OTMEnvironment that
supplies the most used data strings throughout the application. We all
fix a quoting issue on date strings which had an extra ' at the end.