Allow long choices to span multiple lines.
Some of the choices for Stewarships and Alerts are long strings. For
example: "Mulching, adding compost, or amending soil." These don't fix
on a single line of a tableview cell. Here we allow the line to wrap.
This might need to be revisited if there are even longer lines, which
would require actually increasing the size of the cell. But it is a fine
solition for the data we currently have and expect in the near future.

Fixes issue #194.