Fix crash when the address string is Null.
Due to some clean-up associated with work done for Stewardship and Alert
searching I pulled out a check to see if an NSString was actually an
NSNull object. This shouldn't ever really happen so it seemed like the
code could have been cleaned up. However due to the way we handle JSON
responses from the API we actually did have a NSNull object sitting in
string space which caused a crash. The compiler doesn't like seeing
objects of different types in direct comparison. So instead we check the
type of the object in the dictionary directly and only if it is not
NSNull do we pull it out and store it on the string.

Fixes issue #197.