Provide a no data message cell.
When there are no stewardship or alert entries, we were just showing a block of
gray with no lines between stewardship, alerts, or ecosystem benefits.
Instead we add a cell that provides a no data message to the user when
on the tree detail page and not in edit mode.

Fixes issue #196.