Get search radius from API rather than plist.
We had been getting the radius for location and address searches from a
hard coded value it the plist. This was a problem in that the number was
too small for some instances. Better is to calculate the radius from the
bounding box. This was implemented in the API since most of the
geoprocessing is handled there anyway. As a result we only need to grab
the proper key and set it in the app.