Reset instance string if there is a server error.
The way we were handling errors caused us to get into an infinite loop
of rejection of there was a problem with the instance string and the
server responded with an error. This could happen if a user used the app
on an instance and that instance later expired. Reopening the app would
result in a "server error" loop that could only be solved by crashing
out of the app. This resets the instance string if such a failure
occurs. On the cloud app this will result in having to choose another
instance.