Fix location services on iOS8.
In order to use locations on iOS8 we must specify if we intend to use them at
all times or just while the user is actually active in the app. We must specify
a key in our info.plist "NSLocationWhenInUseUsageDescription" or all location
requests are ignored. Prior to calling startUpdatingLocation we must also call
to get the proper authorization. This is not available below iOS8 so we must
test for existence of the method before using it since we are going to target
both 7 and 8. Use of the key in the info plist will prevent the uncomfortable
warning that the app would like to use your location when at all times. Further
reference at:
https://developer.apple.com/library/prerelease/iOS/documentation/CoreLocation/Reference/CLLocationManager_Class/index.html#//apple_ref/occ/instm/CLLocationManager/requestWhenInUseAuthorization