Fix API requests from non-US locales.
I ran across this problem while testing other features. The device was
set to use UK English in non-24 hour time. As a result the date format
that was created and passed to the api was incorrect and caused the
application to fail when loading initial data. This did not happen when
the device was set to non-24 hour time with US English. Since users will
never see this string, we now hard code it to a known good format prior
to formatting it for the API.