Allow editing and saving numeric UDFs
The application was crashing attempting to convert saved numeric UDF
values from numbers to strings, which occurs when showing the edit form
with an integer UDF that has a saved value.

Checking the data type of the UDF and creating a numeric formatter for
the field when it has the ``int`` or ``float`` type activates the
existing numeric field handling, fixing the issue.

The species field has a data type ``int`` (the database id of the selected
species) but has highly customized field handling. I exclude it by name
to prevent potential negative side effects of having ``fmt`` be non-nil
on the species field object.