Support multichoice fields
Editing single choice and multichoice fields have a similar workflow so,
by adding a new `isMulti` property, the existing
`OTMChoicesDetailCellRenderer` can conditionally handle both types of
fields.

When stewardship support was added, the `OTMDetailCellRenderer` base
class used `isKindOfClass:[NSArray class]` to detect a collection UDF
field. Multichoice fields are also stored as arrays, requiring that we
refactor this logic.

The visible difference between the single select and multiselect mode is
subtle, so I added a header to the table view with and instructive
message.

The `stringForValue:` method is repeated in both
`OTMChoicesDetailCellRenderer` and `OTMEditChoicesDetailCellRenderer`
but it is so simple that trying to DRY it out would be more complicated.