Update location handling in nearby trees view
Following the design of the revised `OTMMapViewController` I updated
`OTMNearbyTreesViewController` to use the shared
`CLLocationManager`. The nearby view sets a distance filter on the
`CLLocationManager` so I updated the other view controllers to remove
the filter when they appear.

I also increased the accuracy and decreased minimum distance traveled
before update to improve the accuracy of the results displayed in the
nearby table view.

This was the last view controller to use `OTMLocationManager`, so I
removed it from the project.