Stop location updates on About and Profile views.
These views do not need the user's location so we save the device
battery by turning off location updates. We can not depend on order of
viewWillAppear: and viewWillDisappear: calls when transitioning between
views. We would prefer to put these stopUpdatingLocation calls in the
viewWillDisappear: method of the view controllers that need location
data, but we can't.