Change property attributes on OTMPreferences
The app was crashing with an `EXC_BAD_ACCESS` exception when the app
delegate's `applicationDidEnterBackground:` method attempted to call
`save` on `OTMPreferences`. Setting a breakpoint on unhandled exceptions
revealed that the value of the `instance` property was invalid. It looks
like this is the result of using the `assign` attribute, which should
only be used on primitive types (int, BOOL, etc.).

I am not able to reproduce the exception after replacing `assign` with
our app-wide convention for string properties, `nonatomic` and
`strong`.