Show plot add/update validation messages
An updated version of the API properly return a 400 status code with plot
create/update validation messages. This new code parses the JSON error response
and shows the first global and field-level messages.

A more complete solution would mirror the web app and show the individual
validation errors next to the fields. This is a minimum viable solution that
provides some guidance on how to deal with a save failure rather than showing a
generic "sorry" message.