Exclude readonly when saving tree/plot.
`readonly` is a special field that has not been implemented in the backend.
Editing an empty planting site was incorrectly creating a tree because the app
was sending "false" for the value of `tree.readonly`. The simplest way to handle
this was to simply filter out the `readonly` field before PUTing/POSTing to the
API.