Update date field handling
Before this commit setting the value of a date field to the current date
required changing the date picker to some other value and then changing it back
to the current date. This commit fixes that workflow by setting the date field
value on focus and adding a toolbar with a "Clear" button to remove the value if
it was set accidentally.

To properly clear a previously saved date value required changing the
implementation of `updateDictWithValueFromCell:` to set an `NSNull` in the
dictionary rather than omitting the key from the dictionary entirely.