Fix the workflow when a tree is created but photo saving fails
Previous to these changes a failure to save an image put the application into an
inconsistent state. After showing an error message, the map page was still in
the "add tree" mode with the movable marker appearing on the map.

These changes address that problem by calling slightly different versions the
same post-save actions. We have added a separate and more helpful error message
in the case when saving the photo of a new tree fails vs. when adding a photo to
an existing tree fails.

We have also implemented `viewController:addedTree:` in
`OTMMapViewController.m`. That method was defined in the `.h` file but the
implementation was mistakenly omitted.