Update CocoaPod files and project settings to work with latest
When building with newer versions of CocoaPods we were running into library
search failures. This commit includes updates to the podfile/lock made by
CocoaPods 1.5.3 as well as projects configuration updates that ensure our
dependent libraries are built and located properly.

We also disabled `Parallelize Build` in the build scheme for the application to
avoid a linker error.